Try allowing non-RootTagged HugrMut. Requires duplicating (checking version of) replace_op
